# app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Form Builder API",
        "description": "API for Form Builder application with authentication system",
        "contact": {
            "name": "API Support",
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8085",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/forms": {
            "get": {
                "tags": [
                    "Forms"
                ],
                "summary": "Get all forms for authenticated user",
                "description": "Retrieve a list of all forms belonging to the authenticated user",
                "operationId": "getForms",
                "responses": {
                    "200": {
                        "description": "Forms retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Form"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Forms retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to retrieve forms"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Forms"
                ],
                "summary": "Create a new form",
                "description": "Create a new form for the authenticated user",
                "operationId": "createForm",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "configuration"
                                ],
                                "properties": {
                                    "name": {
                                        "required": [
                                            "en",
                                            "de"
                                        ],
                                        "properties": {
                                            "en": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Contact Form"
                                            },
                                            "de": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Kontaktformular"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "description": {
                                        "properties": {
                                            "en": {
                                                "type": "string",
                                                "maxLength": 1000,
                                                "example": "A contact form for customer inquiries"
                                            },
                                            "de": {
                                                "type": "string",
                                                "maxLength": 1000,
                                                "example": "Ein Kontaktformular für Kundenanfragen"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "configuration": {
                                        "required": [
                                            "locales"
                                        ],
                                        "properties": {
                                            "locales": {
                                                "type": "array",
                                                "items": {
                                                    "type": "string",
                                                    "enum": [
                                                        "en",
                                                        "de",
                                                        "it",
                                                        "fr"
                                                    ]
                                                },
                                                "example": [
                                                    "en",
                                                    "de"
                                                ]
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Form created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Form"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Form created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed to create form"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Form": {
                "title": "Form",
                "description": "Form model",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440000"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "550e8400-e29b-41d4-a716-446655440001"
                    },
                    "name": {
                        "description": "Form name in multiple languages",
                        "properties": {
                            "en": {
                                "type": "string",
                                "example": "Contact Form"
                            },
                            "de": {
                                "type": "string",
                                "example": "Kontaktformular"
                            }
                        },
                        "type": "object"
                    },
                    "description": {
                        "description": "Form description in multiple languages",
                        "properties": {
                            "en": {
                                "type": "string",
                                "example": "A contact form for customer inquiries"
                            },
                            "de": {
                                "type": "string",
                                "example": "Ein Kontaktformular für Kundenanfragen"
                            }
                        },
                        "type": "object"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "configuration": {
                        "description": "Form configuration including locales",
                        "properties": {
                            "locales": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "enum": [
                                        "en",
                                        "de",
                                        "it",
                                        "fr"
                                    ]
                                },
                                "example": [
                                    "en",
                                    "de"
                                ]
                            }
                        },
                        "type": "object"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fields": {
                        "description": "Form fields",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Field"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}